+++ project/src/test/java/es/ubu/lsi/avrela/apm/adapter/github/GitHubApmClientTest.java
@@ -5,7 +5,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import feign.Logger.Level;
-import java.time.LocalDateTime;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -32,10 +31,7 @@ class GitHubApmClientTest {
    * Repository.
    */
   private final String repo = "go-bees";
-  /**
-   * Branch.
-   */
-  private final String branch = "master";
+
   /**
    * Milestone.
    */
@@ -51,45 +47,7 @@ class GitHubApmClientTest {
   }
 
 
-  @Nested
-  @DisplayName("Given a GitHub repository")
-  public class GitHubCommitsTest {
 
-    @Nested
-    @DisplayName("When repository has commits")
-    class GitHubRepositoryWithCommits {
-
-      @Test
-      @DisplayName("Then commits should be fetched")
-      void commitsShouldBeRetrieved() {
-        var commits = gitHubApmClient.findCommits(owner, repo, branch, null, LocalDateTime.now(), 1,
-            1);
-
-        assertNotNull(commits, "Commit list must be none null.");
-        assertTrue(commits.size() > 0, "Commit list length must be greater than zero");
-      }
-
-      @Test
-      @DisplayName("Then commit relevant info should be fetched")
-      void commitsInfoShouldBeComplete() {
-        var commits = gitHubApmClient.findCommits(owner, repo, branch, null, LocalDateTime.now(), 1,
-            100);
-
-        assertAll("Verify relevant info is present",
-            () -> assertTrue(commits.stream().anyMatch(commit -> commit.getSha() != null),
-                "SHA must be retrieved"),
-            () -> assertTrue(
-                commits.stream().anyMatch(commit -> commit.getData().getMessage() != null),
-                "Message must be retrieved"),
-            () -> assertTrue(
-                commits.stream().anyMatch(commit -> commit.getData().getAuthor().getName() != null),
-                "Author name must be retrieved"),
-            () -> assertTrue(
-                commits.stream().anyMatch(commit -> commit.getData().getAuthor().getDate() != null),
-                "Date must be retrieved"));
-      }
-    }
-  }
 
   @Nested
   @DisplayName("Given a GitHub repository")
